+++ core/components/toggletvset/lexicon/en/default.inc.php
@@ -6,9 +6,5 @@
  * @subpackage lexicon
  */
 
-$_lang['setting_toggletvset.debug'] = 'Debug';
-$_lang['setting_toggletvset.debug_desc'] = 'Log debug information in the MODX error log.';
-$_lang['setting_toggletvset.toggletvs'] = 'Toggling Template Variables';
-$_lang['setting_toggletvset.toggletvs_desc'] = 'Comma separated list of template variable ids that should toggle the visibility of other template variables.';
-$_lang['setting_toggletvset.toggletvs_clearhidden'] = 'Clear Hidden Template Variables';
-$_lang['setting_toggletvset.toggletvs_clearhidden_desc'] = 'Clear template variables that are hidden by ToggleTVSet.';
+$_lang['toggletvset'] = 'Toggle TV';
+$_lang['toggletvset.package'] = 'ToggleTVSet';