+++ core/components/toggletvset/lexicon/en/setting.inc.php
@@ -1,0 +1,14 @@
+<?php
+/**
+ * Setting Lexicon Entries for ToggleTVSet
+ *
+ * @package toggletvset
+ * @subpackage lexicon
+ */
+
+$_lang['setting_toggletvset.debug'] = 'Debug';
+$_lang['setting_toggletvset.debug_desc'] = 'Log debug information in the MODX error log.';
+$_lang['setting_toggletvset.toggletvs'] = 'Toggling Template Variables';
+$_lang['setting_toggletvset.toggletvs_desc'] = 'Comma separated list of template variable ids that should toggle the visibility of other template variables.';
+$_lang['setting_toggletvset.toggletvs_clearhidden'] = 'Clear Hidden Template Variables';
+$_lang['setting_toggletvset.toggletvs_clearhidden_desc'] = 'Clear template variables that are hidden by ToggleTVSet.';